requirements.txt
Django==1.6.5
pinax-theme-bootstrap==5.4.0
django-user-accounts==1.0c9
metron==1.3.1
eventlog==0.8.0
django-jsonfield==0.9.13
django-localflavor==1.0
stripe==1.18.0